P3/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conflict-aware-rag"
version = "0.1.0"
description = "Conflict-aware retrieval layer MVP for evidence-centric RAG experiments."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "httpx>=0.27.0",
  "numpy>=1.26.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "qdrant-client>=1.9.0",
  "rank-bm25>=0.2.2",
  "sentence-transformers>=3.0.0",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
